% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtproj-package.R
\docType{package}
\name{dtproj-package}
\alias{dtproj}
\alias{dtproj-package}
\title{dtproj: Create Data Team Projects In A Flash!}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Patrick Anker \email{psanker@nyu.edu} (\href{https://orcid.org/0000-0003-2302-0445}{ORCID})

Other contributors:
\itemize{
  \item Global TIES for Children [copyright holder]
}

}
\keyword{internal}
