% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create a Data Team Pipeline project}
\usage{
init(path = ".", force = FALSE)
}
\arguments{
\item{path}{Where the project should be created. Defaults to
the current working directory.}

\item{force}{If content already exists in \code{path}, removes it for
creation of project architecture. \strong{Be careful!!} Use only
when absolutely needed.}
}
\value{
The path of the project, invisibly
}
\description{
Sets up the necessary infrastructure to get a pipeline spun up
real quick!
}
