% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtproj-package.R
\docType{package}
\name{dtproj-package}
\alias{dtproj}
\alias{dtproj-package}
\title{dtproj: Create Data Team Projects In A Flash!}
\description{
Sets up data projects based on a checklist,
    heavily drawing on the templating features from usethis
    and whisker. Primarily for internal purposes, but could
    be useful more broadly.
}
\author{
\strong{Maintainer}: Patrick Anker \email{psanker@nyu.edu} (\href{https://orcid.org/0000-0003-2302-0445}{ORCID})

Other contributors:
\itemize{
  \item Global TIES for Children [copyright holder]
}

}
\keyword{internal}
